pyproject.toml
[project]
name = "openvoice-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.2.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-core>=0.2.10",
    "litellm[proxy]>=1.79.2",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "piper-tts>=1.4.1",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "silero-vad>=6.2.1",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "onnxruntime-gpu>=1.24.2",
    "langchain-groq>=1.1.2",
    "qwen-asr>=0.0.1",
    "transformers>=4.45.0",
    "huggingface-hub>=0.34.0,<1.0",
    "langgraph-checkpoint-mongodb>=0.3.1",
    "pymongo>=4.11.1",
    "fastapi>=0.115.0",
]
